% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acast.R
\name{acast}
\alias{acast}
\title{Serialize a list}
\usage{
acast(ri, ci, values, nrow, ncol, fill)
}
\arguments{
\item{ri}{0 based row indice}

\item{ci}{0 based column indice}

\item{values}{matrix values}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{fill}{value for missing values}
}
\value{
matrix
}
\description{
Write TSON specification binary-encoded format to a connection.
}
\examples{
## Example
}
